requirements.txt
PySide6
python-can
pynput
RPi.GPIO
spidev